+++ tests/chunk-generation-queue.test.ts
@@ -27,8 +27,8 @@ class FakeChunk {
     return new Float32Array([1, 1, 1]);
   }
 
-  public cubeAmbientOcclusion(): Float32Array {
-    return new Float32Array(24).fill(3);
+  public cubeAmbientOcclusion(): Uint8Array {
+    return new Uint8Array(24).fill(3);
   }
 
   public surfaceHeights(): Uint8Array {
